+++ src/components/Point/Point.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import tokens from '@shopify/polaris-tokens';
 import {Color} from 'types';
 
-import {getColorValue} from '../../../../../../utilities';
+import {getColorValue} from '../../utilities';
 
 interface Props {
   active: boolean;