+++ src/components/StackedAreaChart/components/index.ts
@@ -1,0 +1,3 @@
+export {Legend} from './Legend';
+export {Tooltip} from './Tooltip';
+export {StackedAreas} from './StackedAreas';